+++ CHANGELOG.md
@@ -1,6 +1,9 @@
+v0.6.0:
+  * Update to Aravis 0.8.5.
+
 v0.5.0:
   * Disable "bayer" feature by default.
-  * Fix use of MaybeUnit for latest nightly.
+  * Fix use of `MaybeUnit` for latest nightly.
   * Add fallbacks for unstable features so the crate work on stable.
 
 v0.4.1:
@@ -20,10 +23,10 @@ v0.3.0:
   * Target officially released Aravis 0.7.5.
 
 v0.2.1:
-  * Target aravis 0.7.5.
+  * Target Aravis 0.7.5.
 
 aravis v0.2.0 and aravis-utils v0.2.0:
-  * Update to aravis v0.7.99 (current git master)
+  * Update to Aravis v0.7.99 (current git master)
 
 aravis v0.1.0 and aravis-utils v0.1.2:
   * Update dependencies.